# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Microservice para integração de cadastro/consulta/edição/remoção de empresas na Focus NFe (v2).",
        "title": "Focus Integration Service",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/v2/cnpjs/{cnpj}": {
            "get": {
                "description": "Proxy para Focus: GET /v2/cnpjs/{cnpj}. Informe 14 dígitos (somente números).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "CNPJs"
                ],
                "summary": "Consulta cadastro de CNPJ",
                "parameters": [
                    {
                        "type": "string",
                        "description": "CNPJ (14 dígitos, somente números)",
                        "name": "cnpj",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.FocusCnpjResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.RawPayload"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.RawPayload"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handler.RawPayload"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/handler.RawPayload"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.RawPayload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.RawPayload": {
            "type": "object",
            "additionalProperties": {}
        },
        "model.FocusCnpjEndereco": {
            "type": "object",
            "properties": {
                "bairro": {
                    "type": "string"
                },
                "cep": {
                    "type": "string"
                },
                "codigo_ibge": {
                    "type": "string"
                },
                "codigo_municipio": {
                    "type": "string"
                },
                "codigo_siafi": {
                    "type": "string"
                },
                "complemento": {
                    "type": "string"
                },
                "logradouro": {
                    "type": "string"
                },
                "nome_municipio": {
                    "type": "string"
                },
                "numero": {
                    "type": "string"
                },
                "uf": {
                    "type": "string"
                }
            }
        },
        "model.FocusCnpjResponse": {
            "type": "object",
            "properties": {
                "cnae_principal": {
                    "type": "string"
                },
                "cnpj": {
                    "type": "string"
                },
                "endereco": {
                    "$ref": "#/definitions/model.FocusCnpjEndereco"
                },
                "optante_mei": {
                    "type": "boolean"
                },
                "optante_simples_nacional": {
                    "type": "boolean"
                },
                "razao_social": {
                    "type": "string"
                },
                "situacao_cadastral": {
                    "type": "string"
                }
            }
        },
...
    }
}